+++ src/pysight/main_multiscaler_readout.py
@@ -6,7 +6,6 @@ __author__: Hagai
 from pysight.tkinter_gui_multiscaler import GUIApp
 from pysight.tkinter_gui_multiscaler import verify_gui_input
 from pysight.output_tools import generate_output_list
-from dask import delayed
 
 
 def main_data_readout(gui):
@@ -21,7 +20,7 @@ def main_data_readout(gui):
     # Open file and find the needed parameters
     timepatch = fileIO_tools.get_timepatch(gui.filename.get())
     if timepatch == '3':
-        raise NotImplementedError('Timepatch value "3" is currently not supported. Please message package owner.')
+        raise NotImplementedError('Timepatch value "3" is currently not supported. Please message the package owner.')
     data_range = fileIO_tools.get_range(gui.filename.get())
 
     start_of_data_pos = fileIO_tools.get_start_pos(gui.filename.get())
@@ -31,24 +30,24 @@ def main_data_readout(gui):
 
     # Read the file into a variable
     print('Reading file {}'.format(gui.filename.get()))
-    dict_of_slices = delayed(timepatch_switch.ChoiceManager().process)(timepatch)
-    data = delayed(fileIO_tools.read_lst)(filename=gui.filename.get(), start_of_data_pos=start_of_data_pos,
+    dict_of_slices = timepatch_switch.ChoiceManager().process(timepatch)
+    data = fileIO_tools.read_lst(filename=gui.filename.get(), start_of_data_pos=start_of_data_pos,
                                  timepatch=timepatch)
 
     if gui.debug.get() == 0:
         print('File read. Sorting the file according to timepatch...')
-        df_after_timepatch = delayed(lst_tools.tabulate_input)(data=data, dict_of_slices=dict_of_slices, data_range=data_range,
+        df_after_timepatch = lst_tools.tabulate_input(data=data, dict_of_slices=dict_of_slices, data_range=data_range,
                                                       input_channels=dict_of_input_channels)
 
     elif gui.debug.get() == 1:
         print('[DEBUG] File read. Sorting the file according to timepatch...')
-        df_after_timepatch = delayed(lst_tools.tabulate_input)(data=data[:1e6], dict_of_slices=dict_of_slices,
+        df_after_timepatch = lst_tools.tabulate_input(data=data[:1e6], dict_of_slices=dict_of_slices,
                                                       data_range=data_range, input_channels=dict_of_input_channels)
 
     print('Sorted dataframe created. Starting setting the proper data channel distribution...')
 
     # Assign the proper channels to their data and function
-    dict_of_data = delayed(lst_tools.determine_data_channels)(df=df_after_timepatch,
+    dict_of_data = lst_tools.determine_data_channels(df=df_after_timepatch,
                                                      dict_of_inputs=dict_of_input_channels,
                                                      num_of_frames=int(gui.num_of_frames.get()),
                                                      x_pixels=int(gui.x_pixels.get()),
@@ -57,19 +56,18 @@ def main_data_readout(gui):
                                                      binwidth=float(gui.binwidth.get()))
     print('Channels of events found. Allocating photons to their frames and lines...')
 
-    df_allocated = delayed(lst_tools.allocate_photons)(dict_of_data=dict_of_data, gui=gui)
+    df_allocated = lst_tools.allocate_photons(dict_of_data=dict_of_data, gui=gui)
     print('Relative times calculated. Creating Movie object...')
 
     # Create a movie object
-    final_movie = delayed(class_defs.Movie)(data=df_allocated, x_pixels=int(gui.x_pixels.get()),
+    final_movie = class_defs.Movie(data=df_allocated, x_pixels=int(gui.x_pixels.get()),
                                    y_pixels=int(gui.y_pixels.get()), z_pixels=int(gui.z_pixels.get()),
                                    reprate=float(gui.reprate.get()), name=gui.filename.get(),
                                    binwidth=float(gui.binwidth.get()))
 
     # Find out what the user wanted and output it
     print('======================================================= \nOutputs:\n--------')
-    comp_items = final_movie.compute()
-    output_list = generate_output_list(comp_items, gui)
+    output_list = generate_output_list(final_movie, gui)
     return df_allocated, final_movie, output_list